fix stuff broken by EventArgs changes, move ConnectionEventArgs
